requirements.txt
anyio==3.7.0
asttokens==2.2.1
asyncstdlib==3.10.8
autograd==1.6.2
autograd-gamma==0.5.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.100.0
fonttools==4.42.1
future==0.18.3
h11==0.14.0
httpcore==0.17.3
httptools==0.5.0
httpx==0.24.1
idna==3.4
ipykernel==6.23.2
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
mplcursors==0.5.2
nest-asyncio==1.5.6
numpy==1.25.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.5.3
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.9
Pygments==2.15.1
pyodbc==4.0.39
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
reliability==0.8.12
requests==2.31.0
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
